requirements.txt
asgiref==3.5.2
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
bottle==0.12.23
certifi==2022.6.15
cffi==1.14.5
channels==3.0.3
charset-normalizer==2.1.1
cmake==3.24.1
constantly==15.1.0
cryptography==3.4.6
cycler==0.11.0
daphne==3.0.1
dj-database-url==1.0.0
Django==4.1
django-environ==0.9.0
django-username-email==2.5.3
djangorestframework==3.13.1
dlib==19.24.0
environ==1.0
fonttools==4.37.1
hyperlink==21.0.0
idna==3.1
ikpy==3.0.1
incremental==21.3.0
joblib==1.1.0
kiwisolver==1.4.4
matplotlib==3.5.3
mpmath==1.2.1
numpy==1.23.2
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
pyparsing==3.0.9
pypot==5.0.2
pyserial==3.5
python-dateutil==2.8.2
python-decouple==3.4
pytz==2021.1
requests==2.28.1
scikit-learn==1.1.2
scipy==1.9.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
sympy==1.11
threadpoolctl==3.1.0
tornado==6.2
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
tzdata==2022.2
urllib3==1.26.12
wget==3.2
zope.interface==5.2.0
